requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.3
django-admin-shortcuts==1.2.7
django-ckeditor==5.3.0
django-debug-toolbar==1.8
django-js-asset==0.1.1
django-recaptcha2==1.0.3
django-seo2==1.0.1
idna==2.5
olefile==0.44
Pillow==4.2.0
pytz==2017.2
requests==2.18.1
singlemodeladmin==0.7
six==1.10.0
sqlparse==0.2.3
urllib3==1.21.1
